requirements.txt
ipykernel
pyyaml
requests
pandas